flush log file after each event logged
It may not be good for performance to flush  data  to  disk  after  each
event is logged, but  this  is  necessary  to  ensure  that,  if  server
crashes, the logs are indeed written  into  disk.  Someone  could  argue
about some tradeoff between performance and  reliability,  but  in  this
case we have chosen to really sacrifice performance (which in this case,
isn't a big deal) to get the reliability we want.